create vote when publishing a proposal
